need to be executable

+++ malloc_tester.sh
@@ -140,6 +140,11 @@ if [ ! -f "$EJECUTABLE" ]; then
   echo "The executable does not exist in the specified location."
   exit 1
 fi
+file_type=$(file -b "$EXECUTABLE_PATH")
+if [[ $file_type != *executable* ]]; then
+	echo "Invalid file."
+	exit 1
+fi
 
 rm -f $ADDRESSFILE
 touch $ADDRESSFILE
